# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vicinopoli",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/push/subscriptions": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Subscribe",
        "description": "Register the device's subscription, keyed by the Web Push endpoint.\n\nThe address is geocoded like a post's, but only the geohash cell centre is\npersisted \u2014 re-subscribing from a new address moves the cell.",
        "operationId": "subscribe_api_push_subscriptions_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Unsubscribe",
        "description": "Remove the subscription for the given endpoint (opt-out).",
        "operationId": "unsubscribe_api_push_subscriptions_delete",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscriptionDelete"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushSubscriptionCreate": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Endpoint"
          },
          "p256dh": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "P256Dh"
          },
          "auth": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "Auth"
          },
          "address": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "p256dh",
          "auth",
          "address"
        ],
        "title": "PushSubscriptionCreate",
        "description": "Register a Web Push subscription for the calling device.\n\n``address`` is sent on the wire for geocoding (as post creation does) but\nonly a geohash cell centre is persisted \u2014 never the address or an exact\npoint."
      },
      "PushSubscriptionDelete": {
        "properties": {
          "endpoint": {
            "type": "string",
            "maxLength": 512,
            "minLength": 8,
            "title": "Endpoint"
          }
        },
        "type": "object",
        "required": [
          "endpoint"
        ],
        "title": "PushSubscriptionDelete",
        "description": "Remove a previously registered subscription by its endpoint."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}